schemas + model + url encode + url normalization

+++ app/models/comparison.py
@@ -1,10 +1,10 @@
 from sqlalchemy import Column,Integer,String,Float,DateTime,ForeignKey
 from sqlalchemy.sql import func
-from app.core.db import Base
+from app.core.database import Base
 
 class ClaimComparison(Base):
     __tablename__ = 'claim_comparsion'
-    id = Column(Integer(255),primary_key=True,index=True)
+    id = Column(Integer,primary_key=True,index=True)
 
     claim_a_id = Column(Integer,ForeignKey('claims.id',ondelete='CASCADE'))
     claim_b_id = Column(Integer,ForeignKey('claims.id',ondelete='CASCADE'))
